SinglyLinkedList : insertAllAtEnd -- inserts all the nodes from the head by doing a deep copy of the nodes and adding them to the list
Partition : This program partitions a linkedlist based on a value specified by creating 2 sublists and then merging them together.

Time Complexity - O(2N) ~ O(N) -- since we loop through the main list and loop through the left and right list to construct the main list. It is not necessary to merge the lists this way , they can also be merged by keeping track of the last node of the left list and setting the rightHead as the next pointer to this node.

Space Complexity - O(2N) ~ O(N) since we create 2 new linked lists (deep copy is done both times)